requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
python-dotenv
uuid
PyJWT
werkzeug